+++ examples/src/bin/parl_io_rx.rs
@@ -14,7 +14,8 @@ use esp_hal::{
     delay::Delay,
     dma::{Dma, DmaPriority},
     dma_buffers,
-    parl_io::{no_clk_pin, BitPackOrder, ParlIoRxOnly, RxFourBits},
+    gpio::NoPin,
+    parl_io::{BitPackOrder, ParlIoRxOnly, RxFourBits},
     prelude::*,
 };
 use esp_println::println;
@@ -34,6 +35,7 @@ fn main() -> ! {
         peripherals.GPIO3,
         peripherals.GPIO4,
     );
+    let mut rx_clk_pin = NoPin;
 
     let parl_io = ParlIoRxOnly::new(
         peripherals.PARL_IO,
@@ -45,7 +47,12 @@ fn main() -> ! {
 
     let mut parl_io_rx = parl_io
         .rx
-        .with_config(&mut rx_pins, no_clk_pin(), BitPackOrder::Msb, Some(0xfff))
+        .with_config(
+            &mut rx_pins,
+            &mut rx_clk_pin,
+            BitPackOrder::Msb,
+            Some(0xfff),
+        )
         .unwrap();
 
     let mut buffer = rx_buffer;